Stable: R2.1,2.2,2.3,2.4,2.5,2.6,2.7,2.8,2.9,2.11,
The following items are incomplete and need a few changes

R2.6) R2.7)
Half done; word is showing up, but different shape is not happening
right now.

R2.8)
Remove is removing the entire word. For example, if apple and app are
typed, removing app will remove apple as well, since they share a
common root. This is half done.